+++ bin/mutator-graphite.rb
@@ -9,10 +9,10 @@
 #   (so you don't need to use OnlyCheckOutput) and it also modifies
 #   the format of the hostname in the output if present.
 #
-#   Note however that using this mutator as an mutator command can be very
+#   Note however that using this mutator as a mutator command can be very
 #   expensive, as Sensu has to spawn a new Ruby process to launch this script
-#   for each result of a metrics check. Consider instead to produce the correct
-#   metric names from your plugin and send them directly to Graphite via the
+#   for each result of a metrics check. Consider instead producing the correct
+#   metric names from your plugin and sending them directly to Graphite via the
 #   socket handler.
 #   See https://groups.google.com/d/msg/sensu-users/1hkRSvL48ck/8Dhl98lR24kJ
 #   for more information.
@@ -20,7 +20,7 @@
 # OUTPUT:
 #   Sensu event output with all dots changed to underlines in host name
 #   If -r or --reverse parameter given script put hostname in reverse order
-#   for better graphite tree view
+#   for better Graphite tree view
 #
 # PLATFORM:
 #   all
@@ -40,5 +40,5 @@ event = JSON.parse(STDIN.read, symbolize_names: true)
 if ARGV[0] == '-r' || ARGV[0] == '--reverse'
   puts event[:check][:output].gsub(event[:client][:name], event[:client][:name].split('.').reverse.join('.'))
 else
-  puts event[:check][:output].gsub(event[:client][:name], event[:client][:name].gsub('.', '_'))
+  puts event[:check][:output].gsub(event[:client][:name], event[:client][:name].tr('.', '_'))
 end